✨split server/client component to render dummy random text codes
